backend/requirements.txt
fastapi>=0.95
uvicorn[standard]>=0.20
pydantic>=1.10
sqlalchemy>=1.4
asyncpg>=0.27
pgvector==0.4.2
openai>=1.0
httpx>=0.24
python-dotenv>=1.0
pytest>=7.0
pytest-asyncio>=0.21
respx>=0.20
